refactor : TeamMemberResponseUtils로 구조 개선
