+++ src/main/java/site/codemonster/comon/domain/team/controller/TeamController.java
@@ -11,6 +11,7 @@ import site.codemonster.comon.domain.team.service.TeamService;
 import site.codemonster.comon.domain.team.utils.TeamResponseUtils;
 import site.codemonster.comon.domain.teamMember.entity.TeamMember;
 import site.codemonster.comon.domain.teamMember.service.TeamMemberService;
+import site.codemonster.comon.domain.teamMember.utils.TeamMemberResponseUtils;
 import site.codemonster.comon.domain.teamRecruit.entity.TeamRecruit;
 import site.codemonster.comon.domain.teamRecruit.service.TeamRecruitService;
 import site.codemonster.comon.global.error.dto.response.ApiResponse;
@@ -45,6 +46,7 @@ public class TeamController {
     private final MemberService memberService;
     private final TeamRecruitService teamRecruitService;
     private final TeamResponseUtils teamResponseUtils;
+    private final TeamMemberResponseUtils teamMemberResponseUtils;
 
     @PostMapping
     public ResponseEntity<ApiResponse<?>> createTeam(
@@ -323,7 +325,7 @@ public class TeamController {
     ){
         List<TeamMemberResponse> teamMemberResponses = teamMemberService.getTeamMembersByTeamId(teamId, member)
                 .stream()
-                .map(teamMemberService::getTeamMemberResponse)
+                .map(teamMemberResponseUtils::getTeamMemberResponse)
                 .toList();
 
         return ResponseEntity.status(HttpStatus.OK)
